Batch calls to get vault info (#21)
Most of the bot's time today is spent fetching vaults data for every vault registered in the system.
Every 'get_vault_info' call is a multicall with 3 internal calls. We have 800 vaults => 800 separate multicalls
are made at each bot iteration. This is painfully slow (look at all the gaps here https://info.yieldprotocol.com/#/liquidations)

This change batches together multicalls for different vaults, it speeds up fetching the vaults info at least 10x.
We now do 2 levels of batching:
  1. For each vault_id, there are 3 calls made [this is unchanged]
  2. Calls for different vaults are batched together (self.multicall_batch_size is the batch size).

If multicall_batch_size == 10 and we query 42 vaults, we will issue 5 separate Multicall2 calls:
 * first 4 multicalls will have multicall_batch_size * 3 == 40 internal calls
 * last multicall will have 2 * 3 = 6 internal calls

--multicall_batch_size controls the batch size, default value is 1000

We also switch to Multicall2 instead of Multicall (if a single vault in a batch is bad and we can't
fetch its info, we can't fail the entire batch - Multicall2 allows for that)